+++ rpcmock/mock.go
@@ -1,4 +1,4 @@
-package mock
+package rpcmock
 
 import (
 	"reflect"
