+++ rpcmock/mock_test.go
@@ -1,4 +1,4 @@
-package mock
+package rpcmock
 
 import (
 	"errors"
